A continued work of #32

Changes on top of what the original PR improved:
* Using a mempool for avoid 1 allocation in Get.
* Configuration of mempooling is optional. The default behavior is not using mempool.
* Additional table-tests with mempool to compare face-to-face.
* Rollback of the mempool proposed in the `streampb` package.

I went forward and backward with the type of pool to use. My original intention was `SizedBufferPool` but that was quite complicated because the max size of each element in the pool would be the maximum size of `proto.Entry`, which isn't trivial to calculate considering `maxKeySize` and `maxValueSize`.

As a solution, I proposed using `BufferPool` and always see if the current `Buffer` is enough to the requested size. If isn't the case, Grow it appropriately. The pool entries will eventually converge to the worst usual case.

I also considered adding one more parameter to the new `Option`: `MaxTotalMemPoolSize` to set a hard-limit on mempool size, but I felt wasn't quite trivial to use. Each max buff in the pool would be `MaxTotalMemPoolSize/MaxConcurrency`, and if `MaxTotalMemPoolSize` is set too low would be a problem. Moreover, estimating a reasonable value implies trying to guess what is the expected encoded proto size for values. (This may be confusing, I can reword it if isn't that clear).

Results:
```
goos: linux
goarch: amd64
pkg: github.com/prologic/bitcask
BenchmarkGet/128B-8              1000000              1152 ns/op         111.02 MB/s         387 B/op          4 allocs/op
BenchmarkGet/128BWithPool-8      1000000              1189 ns/op         107.59 MB/s         227 B/op          3 allocs/op
BenchmarkGet/256B-8              1000000              1255 ns/op         203.86 MB/s         643 B/op          4 allocs/op
BenchmarkGet/256BWithPool-8      1000000              1246 ns/op         205.33 MB/s         355 B/op          3 allocs/op
BenchmarkGet/512B-8              1000000              1416 ns/op         361.47 MB/s        1187 B/op          4 allocs/op
BenchmarkGet/512BWithPool-8      1000000              1357 ns/op         377.10 MB/s         611 B/op          3 allocs/op
BenchmarkGet/1K-8                1000000              1763 ns/op         580.76 MB/s        2275 B/op          4 allocs/op
BenchmarkGet/1KWithPool-8        1000000              1592 ns/op         642.89 MB/s        1123 B/op          3 allocs/op
BenchmarkGet/2K-8                 500000              2415 ns/op         847.88 MB/s        4451 B/op          4 allocs/op
BenchmarkGet/2KWithPool-8        1000000              2009 ns/op        1019.35 MB/s        2147 B/op          3 allocs/op
BenchmarkGet/4K-8                 300000              3854 ns/op        1062.67 MB/s        9059 B/op          4 allocs/op
BenchmarkGet/4KWithPool-8         500000              3029 ns/op        1352.22 MB/s        4195 B/op          3 allocs/op
BenchmarkGet/8K-8                 200000              6173 ns/op        1326.90 MB/s       17763 B/op          4 allocs/op
BenchmarkGet/8KWithPool-8         300000              4911 ns/op        1668.07 MB/s        8291 B/op          3 allocs/op
BenchmarkGet/16K-8                200000             10697 ns/op        1531.57 MB/s       34915 B/op          4 allocs/op
BenchmarkGet/16KWithPool-8        200000              7763 ns/op        2110.51 MB/s       16483 B/op          3 allocs/op
BenchmarkGet/32K-8                100000             21858 ns/op        1499.11 MB/s       73827 B/op          4 allocs/op
BenchmarkGet/32KWithPool-8        100000             14607 ns/op        2243.25 MB/s       32867 B/op          3 allocs/op
```
Changes in the PR branch are allowed, so feel free to add further commits.