@prologic 2019-08-03
Here are the same results on my 11" Macbook:

```
--- PASS: TestLocking (0.01s)
goos: darwin
goarch: amd64
pkg: github.com/prologic/bitcask
BenchmarkGet/128B-4         	  300000	      4126 ns/op	  31.02 MB/s	     387 B/op	       4 allocs/op
BenchmarkGet/128BWithPool-4 	  300000	      4275 ns/op	  29.94 MB/s	     227 B/op	       3 allocs/op
BenchmarkGet/256B-4         	  300000	      3970 ns/op	  64.47 MB/s	     643 B/op	       4 allocs/op
BenchmarkGet/256BWithPool-4 	  300000	      4570 ns/op	  56.01 MB/s	     355 B/op	       3 allocs/op
BenchmarkGet/512B-4         	  300000	      4843 ns/op	 105.71 MB/s	    1187 B/op	       4 allocs/op
BenchmarkGet/512BWithPool-4 	  200000	      5622 ns/op	  91.07 MB/s	     611 B/op	       3 allocs/op
BenchmarkGet/1K-4           	  300000	      5885 ns/op	 173.99 MB/s	    2275 B/op	       4 allocs/op
BenchmarkGet/1KWithPool-4   	  200000	      5886 ns/op	 173.96 MB/s	    1123 B/op	       3 allocs/op
BenchmarkGet/2K-4           	  200000	      6859 ns/op	 298.56 MB/s	    4451 B/op	       4 allocs/op
BenchmarkGet/2KWithPool-4   	  200000	      6724 ns/op	 304.55 MB/s	    2147 B/op	       3 allocs/op
BenchmarkGet/4K-4           	  200000	      8271 ns/op	 495.21 MB/s	    9059 B/op	       4 allocs/op
BenchmarkGet/4KWithPool-4   	  200000	      7429 ns/op	 551.32 MB/s	    4195 B/op	       3 allocs/op
BenchmarkGet/8K-4           	  200000	     10900 ns/op	 751.53 MB/s	   17763 B/op	       4 allocs/op
BenchmarkGet/8KWithPool-4   	  200000	      9332 ns/op	 877.77 MB/s	    8291 B/op	       3 allocs/op
BenchmarkGet/16K-4          	  100000	     14014 ns/op	1169.07 MB/s	   34915 B/op	       4 allocs/op
BenchmarkGet/16KWithPool-4  	  100000	     12152 ns/op	1348.19 MB/s	   16483 B/op	       3 allocs/op
BenchmarkGet/32K-4          	  100000	     23211 ns/op	1411.70 MB/s	   73827 B/op	       4 allocs/op
BenchmarkGet/32KWithPool-4  	  100000	     18618 ns/op	1759.97 MB/s	   32867 B/op	       3 allocs/op
BenchmarkPut/128B-4         	  100000	     13048 ns/op	   9.81 MB/s	     409 B/op	       6 allocs/op
BenchmarkPut/256B-4         	  100000	     15815 ns/op	  16.19 MB/s	     539 B/op	       6 allocs/op
BenchmarkPut/512B-4         	  100000	     20986 ns/op	  24.40 MB/s	     829 B/op	       6 allocs/op
BenchmarkPut/1K-4           	   50000	     27889 ns/op	  36.72 MB/s	    1411 B/op	       6 allocs/op
BenchmarkPut/2K-4           	   50000	     37723 ns/op	  54.29 MB/s	    2574 B/op	       6 allocs/op
BenchmarkPut/4K-4           	   20000	     72584 ns/op	  56.43 MB/s	    5156 B/op	       6 allocs/op
BenchmarkPut/8K-4           	   10000	    124705 ns/op	  65.69 MB/s	    9814 B/op	       6 allocs/op
BenchmarkPut/16K-4          	   10000	    195204 ns/op	  83.93 MB/s	   18851 B/op	       6 allocs/op
BenchmarkPut/32K-4          	    5000	    350503 ns/op	  93.49 MB/s	   41549 B/op	       7 allocs/op
BenchmarkScan-4             	 1000000	      1943 ns/op	     493 B/op	      25 allocs/op
PASS
ok  	github.com/prologic/bitcask	49.615s
```