@jsign 2019-08-03
This PR in a fork of your branch in the other PR. This PR has the same commits you authored in the other PR. Since I'm not a contributor, I couldn't push new commits to your branch to continue on the other PR.

I rebased to master too.