@prologic 2019-08-03
> Since I'm not a contributor, I couldn't push new commits to your branch to continue on the other PR (or thats what I think since I coudn't push to it)

Oh that can easily be fixed! One moment...