@prologic 2019-08-03
Please feel free to push to direct branches. I have a few minor comments on this PR:

- Why do we not just set `memPool` to `nil` by default and avoid passing in int poiners `*int`?
- The boop library calls the paramter `maxSize` and we're callig it `maxConcurrency`? Should we stick with `maxSize` to be consistent?