@prologic 2019-08-03
Should we also employ the use of a fixed-size `ByteBuffer` for allocating the prefixSize framing used which will just speed up reading/streaming the whole database (_in merge operations_)?

```
prologic@Jamess-MacBook
Sun Aug 04 09:39:51
~/Projects/bitcask
 (buffer_pool2) 0
$ sift -C 3 'make.*byte'
internal/datafile.go-143-
internal/datafile.go-144-	var b []byte
internal/datafile.go-145-	if memPool == nil {
internal/datafile.go:146:		b = make([]byte, size)
internal/datafile.go-147-	} else {
internal/datafile.go-148-		poolSlice := memPool.Get()
internal/datafile.go-149-		if poolSlice.Cap() < int(size) {
--
internal/streampb/stream.go-29-// Encode takes any proto.Message and streams it to the underlying writer.
internal/streampb/stream.go-30-// Messages are framed with a length prefix.
internal/streampb/stream.go-31-func (e *Encoder) Encode(msg proto.Message) (int64, error) {
internal/streampb/stream.go:32:	prefixBuf := make([]byte, prefixSize)
internal/streampb/stream.go-33-
internal/streampb/stream.go-34-	buf, err := proto.Marshal(msg)
internal/streampb/stream.go-35-	if err != nil {
--
internal/streampb/stream.go-67-// Decode takes a proto.Message and unmarshals the next payload in the
internal/streampb/stream.go-68-// underlying io.Reader. It returns an EOF when it's done.
internal/streampb/stream.go-69-func (d *Decoder) Decode(v proto.Message) (int64, error) {
internal/streampb/stream.go:70:	prefixBuf := make([]byte, prefixSize)
internal/streampb/stream.go-71-
internal/streampb/stream.go-72-	_, err := io.ReadFull(d.r, prefixBuf)
internal/streampb/stream.go-73-	if err != nil {
--
internal/streampb/stream.go-76-
internal/streampb/stream.go-77-	n := binary.BigEndian.Uint64(prefixBuf)
internal/streampb/stream.go-78-
internal/streampb/stream.go:79:	buf := make([]byte, n)
internal/streampb/stream.go-80-
internal/streampb/stream.go-81-	idx := uint64(0)
internal/streampb/stream.go-82-	for idx < n {
```