@jsign 2019-08-03
> Please feel free to push to direct branches. I have a few minor comments on this PR:
> 
> * Why do we not just set `memPool` to `nil` by default and avoid passing in int poiners `*int`?
> * The boop library calls the paramter `maxSize` and we're callig it `maxConcurrency`? Should we stick with `maxSize` to be consistent?

* Since the `memPool` is package-wise, if a client creates a mempooled enabled `db`, close its, and then opens it again but want it without pool... this would allow to re-configure the package. This is similar to what I mentioned in the other PR regarding some refactoring. (I don't like package scoped variables for this kind of consequences)
* Maybe, but I think that's polluting the language domain of what the client understands. I guess the `maxSize` concept of `bpool` is an implementation detail which affects the concurrency concept in bitcask. It may be renamed as `maxSize`, but we should clarify a bit more _size of what_ , _why it matters for you_, and _how you should choose this value_ (naming `maxConcurrency` may imply the clarifications, but I guess gives a better hint; not sure about this though).